requirements.txt
fastapi
uvicorn[standard]
pydantic

torch
diffusers
transformers
safetensors
accelerate
Pillow
